chore(rollup): update rollup to the latest version
Relates to #100
